Wait for Google Analytics Cordova to load
Since the tracker initialization is asynchronous, it’s best to wait
until we know it’s complete before calling any of the other methods (at
least, for the WinPhone implementation, this was causing the first
event to get lost).

By moving the tracker start from initialize to load, and waiting until
load is complete before using the tracker, we can guarantee it exists
for the first page view.